refactor: 콘솔, Spark 어플리케이션 관련 기능 삭제 및 클래스명 변경
