--- src/main/java/chess/domain/command/Command.java
@@ -1,9 +1,0 @@
-package chess.domain.command;
-
-public interface Command {
-    void handle(String input);
-
-    boolean isUsable(String input);
-
-    boolean isStatus();
-}
