+++ src/main/java/chess/view/dto/ScoreDto.java
@@ -1,17 +1,11 @@
 package chess.view.dto;
 
-import chess.domain.command.StatusCommand;
 import chess.domain.game.ChessGame;
 
 public class ScoreDto {
 
     private final double whiteScore;
     private final double blackScore;
-
-    public ScoreDto(final StatusCommand statusCommand) {
-        blackScore = statusCommand.getBlackScore();
-        whiteScore = statusCommand.getWhiteScore();
-    }
 
     public ScoreDto(final ChessGame chessGame) {
         blackScore = chessGame.getBlackScore();
