+++ src/main/resources/templates/index.hbs
@@ -20,12 +20,12 @@
     </div>
 
     {{#status}}
-    {{^exist}}
-    <button class="button is-primary is-light chess-btn start" value="start">게임 시작</button>
-    {{/exist}}
-    {{#exist}}
-    <button class="button is-primary is-light chess-btn continue" value="continue">이어하기</button>
-    {{/exist}}
+        {{^exist}}
+            <button class="button is-primary is-light chess-btn start" value="start">게임 시작</button>
+        {{/exist}}
+        {{#exist}}
+            <button class="button is-primary is-light chess-btn continue" value="continue">이어하기</button>
+        {{/exist}}
     {{/status}}
     <button class="button is-warning is-light hidden chess-status-btn">점수 확인</button>
     <button class="button is-danger is-light hidden chess-end-btn">게임 종료</button>
